clean up e2e test
